common/chipset.c: Set AP_IDLE flag when apshutdown
The AP_IDLE flag is not set when apshutdown, the reason AP_IDLE
is not set by default is that apshutdown does a forced shutdown,
so the power sequencing code detects it as a power failure, so
pb_chipset_shutdown() does not set the AP_IDLE flag.

When apshutdown in the recovery mode, the EC will sysjump from RO
to RW, if the device is powered by USB-C, EC will brown out and
reset due to PD hard reset. If the AP_IDLE flag is not set, EC
will boot AP after reset. This patch sets the AP_ILDE flag when
running apshutdown command so the EC will not boot AP after brown
out and reset.

BUG=b:301010349
TEST=Run `apshutdown` command in the EC console to shut down the
     Chromebox device which is in recovery mode and powered by
     USB-C adapter.
LOW_COVERAGE_REASON=command_apshutdown has no tests yet

Change-Id: I7e748ec22eb66a4f341a0717ea76dcebb165bbc2