intel altmode: Add support for TBT and USB4 mode mux
This CL reshuffles Kconfig pd flags related to TBT and USB4 so that they
can be enabled for PD solution. Also adds code to configure mux for TBT
and USB4 mode.
This CL further requires usbc ports device tree for successful
compilation for mtlrvpp_pd which is part of another CL.

BUG=b:295917619
TEST=Able to negotiate TBT and USB4 using CCG8-PD & MTL-RVP

Change-Id: Ida1e96fc7f15deb86d25d53ae2c65fea994b17c5