docs: Add library-compile-flags.md
Add a doc that shows the current compiler flags for use in
generatig a static library.

We can make manual alterations to these flags, but we should
commit this version to doucmnet what the flags were, before
making library specific changes.

This document will be visble at one of the following URLs,
once submitted:
https://crsrc.org/o/src/platform/ec/docs/fingerprint/library-compiler-flags.md
https://chromium.googlesource.com/chromiumos/platform/ec/+/HEAD/docs/fingerprint/library-compiler-flags.md

BUG=b:301146101
TEST=Viewed in Gitiles

Change-Id: Ie0fc7dece82f4d84974a30b4ccd19de20f26a435