cbi: Make CBI_IMAGE_SIZE customizable for EC flash
At present we are using constant value for image size that is 256
bytes. With BCIC data storage in CBI, the size of CBI should be more to
accommodate. We can't increase the default value cause it might overflow
RAM for some devices like Rex, hence we are making it configurable based
on device capacity and need. The default value of image-size property is
256

BUG=b:301200975
TEST=Flashed on Rex, tried setting image-size to 4k for Rex and saw
build failing as expected, tried setting different image-size to EC
flash CBI section against EEPROM and saw build failing as expected

Change-Id: I0686df42aa40a876f088fcfe125bf3a8229505e6