ecusb: Prepare for copying to hdctools
hdctools is pretty strict on pylint enforcement, so clean things up
before copying: Ensure raw string interpretation for doc-string in
light of backslashes in the regex example.

See https://github.com/pylint-dev/pylint/issues/264#issuecomment-163174821
for an example of why other approaches fail, as part of a larger
discussion on the subject.

BUG=none
TEST=none

Change-Id: I3e7bd2f08098937ca26c7b14940090a1bb750978