gsctool: scan for dauntless second if no serial, vid, pid or image
Commands that support both haven and dauntless, e.g. --reboot, benefit
from first scanning for cr50 (current default behavior), but if a haven
chip isn't found, then scan for a dauntless chip.

We do not want this secondary scanning behavior for the case when either
-D is explicitly passed in or if a fw image is explicitly passed in. In
either of those cases, the chip is well known and we should not try a
secondary chip scan.

BUG=none
TEST=verify that cr50 image update will only scan for H1 chip
TEST=verify that ti50 image update will only scan for DT chip
TEST=verify that gsctool --reboot will first scan for H1, then DT
TEST=verify that gstool -D --reboot will only scan for DT chip

Change-Id: I976284d9136dad7ed940863f299e3f4209f9e992