motion_sense: Do not use stale data when checking forced mode
When the EC transitions from RO to RW, the sensors are reset, an ODR
request is sent, but otherwise, existing data are still present.

In particular, the |collection_rate| in increment_sensor_collection() is
still used, but during the sensor init() call, its odr field has been
reset to 0.

So we will assert in motion_sense_read() because ODR is 0 but
|collection_rate| contains the stale value set by RO firmware.

We would crash with:
[2.896482 Lid Accel: MS Done Init type:0x0 range:2]
[2.900263 SENSOR accelgyro_lsm6dsm WHO_AM_I value: 0x6a]
[2.901431 Lid Accel Missed 12 data collections at 1938850 - rate: 80000]
ASSERTION FAILURE
'sensor->drv->get_data_rate(sensor) != 0' in motion_sense_read() at
common/motion_sense.c:553

The message "rate: 80000" is the stale collection_rate in us.
The message "Missed 12 data collections" indicates we could not collect
while transition to RW.

Make sure we set the data needed for forced collection up to date before
using it.

BUG=b:295773669
BRANCH=all
TEST=compile.
Add ODR change request when `accelinit` is invoked.

Change-Id: I8037206a4b6befd4e9d57cda1cc77b6e5a402caf