timer: Prevent frequent error message
Calling usleep from an interrupt, or while interrupts are disabled,
results in a printing a "Sleeping not allowed" warning message. If
usleep is called too frequently this can fill the uart buffer and
result in lost messages. Reduce the frequency of this message to a
frequency of at most 20ms.

BUG=b:291141229
TEST=run_device_tests.py --board helipilot -t rollback_entropy

Change-Id: I6509f0b9d606aad4f2ce4047602bdd720c650030