cec/bitbang: it83xx: Minimise number of GPIO operations in ISR
Currently, the ISR takes a long time - ~140 us on it83xx vs ~25 us on
npcx. With the following change, this can sometimes result in the next
deadline being missed, and sending/receiving messages to fail.

This is mosly due to slow GPIO operations (enable/disable interrupt, set
flags to rising/falling). So instead of updating the GPIO state in every
interrupt, keep interrupts enabled for both edges, and have the
interrupt handler filter out interrupts we're not waiting for.

This reduces the average ISR duration from ~140 us to ~60 us.

BUG=b:287676025
TEST=CEC works reliably on dibbi and kinox

Change-Id: Ia434cda46c6501b4761b420f1f9313ed99e4dc92