cec/bitbang: it83xx: Account for delay before ISR starts
When calculating timeout values, the driver already takes into account
the delay from when the ISR starts to when the timer is programmed.
However, since ITE doesn't have a capture timer and the timing is done
in software, there's an additional delay from when the interrupt
actually occurs to when the ISR starts, and this is not currently
accounted for.

Looking at signal traces, this delay seems to be about 100 us, since
each high and low time is currently ~100 us greater than the nominal
values. So add a hardcoded 100 us delay to account for it.

This fixes an issue with the low time of the ACK bit being out of spec
(b/287676025#comment72 - measured time 1.792 ms, spec max 1.7 ms). The
reason ACK bits are more greatly affected is they involve three
interrupt + state transition cycles (FOLLOWER_ACK_LOW,
FOLLOWER_ACK_VERIFY, FOLLOWER_ACK_FINISH).

BUG=b:287676025
TEST=On dibbi and taranza, ACK bit low time now meets the spec. Other
bit timings are also closer to the nominal value.

Change-Id: I228207422928cfd191df8f2dd3751b4cdffa2793