boxy: Implement power monitoring
If all ports are fully loaded, we could exceed the rating of the adaptor
powering dibbi, causing a brownout. To prevent this we monitor the
total system power and apply throttles if we're close to exceeding the
rating. Specifically, if we're sourcing from the Type-C port, we will
throttle it from 15 W to 7.5 W.

Based on puff with the following changes:
- Only monitor the total system power, not the 5V rail power. On dibbi
  the current limit of the 5V load switch is high enough to handle a
  fully loaded system.
- The only throttle available is the Type-C port power. Type-A power and
  PROCHOT cannot be controlled by the EC on dibbi. However, the CPU
  power will be throttled by PSYS.
- Add hysteresis to the throttling threshold to avoid switching
  frequently due to small power variations.

BUG=b:303008358
TEST=- Connect an E-load to PPVAR_SYS and a device to the Type-C port.
- Increase the E-load power and check the Type-C port power is throttled
  to 7.5 W.
- Decrease the E-load power and check the Type-C power returns to 15 W.

Change-Id: I2253554f6b893d6a57826081d4c6e41082f56f45