zephyr: tasks: disable task creation division-by-zero warning on LLVM
LLVM trigger a harmless division-by-zero on SYS_TIMEOUT_MS, due to an
element of z_tmcvt_64 resulting in a "/ ((1000) / (10000))" in some
specific tick and clock configuration on some platforms.

The resulting value is verfied to be calculated correctly:

(gdb) p _k_thread_data_RWSIG.init_delay
$1 = {ticks = -1}

So the warning is harmless here as the condition never kicks in.

BUG=none
TEST=verified the output structure with gdb

Cq-Depend: chromium:4905538
Change-Id: I47f7d19c2edb49bbc77af32ace854da72efd2098