Cutoff: Check AC disconnect accurately
TCPM resets a PD connection on reboot. This appears as AC-off (via
HOOK_AC_CHANGE) and the battery subsystem triggers cutoff if the refresh
key is being pressed. This is usually immediately canceled by a
subsequent AC-on but there is a chance that a PD re-connection doesn't
finish by the time scheduled cutoff executes.

With this CL, the EC checks the active charge port on
HOOK_POWER_SUPPLY_CHANGE. If it changes from !CHARGE_PORT_NONE to
CHARGE_PORT_NONE, the EC considers all chargers are unplugged and
triggers battery cutoff.

BUG=b:285187778
TEST=Cutoff by Refresh+Unplug using normal & DR chargers.
TEST=No cutoff by keeping Refresh down using normal & DR chargers.

Change-Id: I42bfa750436a29fa1727839cb17f66aa0d5426e4