Cutoff: Detect AC disconnect accurately
the TCPM resets a PD connection on reboot. This appears as AC-off (via
HOOK_AC_CHANGE) and the battery subsystem triggers a cutoff if the
refresh key is being pressed. This is usually immediately canceled by a
subsequent AC-on but there is a chance that a PD re-connection doesn't
finish by the time a scheduled cutoff starts.

With this CL, the EC monitors the active charge port using
HOOK_POWER_SUPPLY_CHANGE. If it changes from !CHARGE_PORT_NONE to
CHARGE_PORT_NONE, the EC considers all chargers are unplugged and
triggers a cutoff.

BUG=b:285187778
TEST=Cutoff by Refresh+Unplug using normal & DR chargers.
TEST=No cutoff by keeping Refresh down using normal & DR chargers.

Change-Id: I42bfa750436a29fa1727839cb17f66aa0d5426e4