Keyboard: Cancel pass-through rows while power button is pressed
While the power button is pressed, Ti50 drives KSO2 to read a refresh
key state but it does not mask KSI2 (or KSI3) forwarded to the EC.
This makes the EC see KSI2 asserted on any column it scans.

With this CL, the keyboard scanner will clear the KSI2 bit if it sees
the PB is pressed. If the power button state changes before and after
reading a column, it re-scans the column. This is expected to work as a
debouncer.

This is a workaround. When Ti50 with crrev/i/4896707 reverted is
released, we should revert this workaround as well.

BUG=b:300418915,b:35575102
TEST=On Rex, type keyboard normally.
TEST=On Rex, press refresh+power to register refresh key as a boot key.
TEST=On Rex, press refresh+power+esc to enter recovery mode.
LOW_COVERAGE_REASON=False positives. See comments on Gerrit.

Change-Id: I547079adc99dcb7e43d2652ff20dd7d38ead1d9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4912024
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Kenny Pan <kennypan@google.com>
Auto-Submit: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Kenny Pan <kennypan@google.com>
Reviewed-by: Wilson Lu <wilsonlu@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7418501
Reviewed-by: Jayvik Desai <jayvik@google.com>
Commit-Queue: Jayvik Desai <jayvik@google.com>
Tested-by: Jayvik Desai <jayvik@google.com>