motion_sense: handle inactive sensors during suspend
In the case we set sensor active mask as SENSOR_ACTIVE_S0 so sensors
are active only during S0; When system suspends, the ODR of sensors
should be set to 0. If AP has set ODR to non-zero value before suspend,
then save it and restore AP ODR value when system resumes.

BUG=b:281038970 b:301017318
TEST=zmake build rex-ish, flash and run, system enter S0ix, check ODR of
sensors are set to 0; system resume, ODR are restored.

Change-Id: I9547acf99f7547f3a09e8217059263008e6b678c