geralt: fix incorrect byte offset of hid-i2c command
The command register is a 16bit little-endian number on byte 2~3 of the
input stream, not on byte 3~4. Fix the incorrect index.

This code worked before since we only care about op-code field, we
luckily read the correct index (lower 4 bits of byte 3). But we still
need to keep the whole command register correct.

BUG=none
TEST=hid-i2c communication between ec and kernel works correctly.

Change-Id: I137cb82ed900bfca300f2354ab7a6524f75dc6d8