ppc/ktu1125: Gracefully disable sink mode
When a charger change event occurs, we sweep all ports to make sure
there are no conflicting conflicting configurations - in particular,
charging from more than one port at a time. When the ktu1125 PPC driver
was told to turn off VBUS sink, it would turn off VBUS when in source
mode instead of treating this request as a no-op.

BUG=b:304611222
TEST=hub no longer turns off when connecting/disconnecting charger

Change-Id: Ida5a3e022a0fc6de9075f68fc474f8dcbe5400a4