docs: Use ChromiumOS SDK toolchain in library-compiler-flags.md
Add note about using ChromiumOS SDK toolchain, since it is too difficult
to use an external toolchain. This is because of the manual setup required
to user newlib and other compiler libraries, correctly.

We would be reinventing the wheel if we gave instructions on how to
build the toolchain from scratch.

BUG=b:301146101
TEST=Viewed in Gitiles

Change-Id: I72f67cdc88dee82a408880240fd9fa38838a4c10