npcx: flash: set_status returns error if hardware WP is eanbled
When the /WP of internal flash is asserted and the SRP0 bit of the
status register 1 is set, writing to status registers should fail.
The flash_set_status function currently doesn't check it and treats it
always successfull. This CL changes the flash_set_status function to
return the error code if the hardware write protect is enabled.

BUG=b:304274482
BRANCH=none
TEST=pass "make buildall"
TEST=pass unit test "test-flash_write_protect"

Change-Id: Id89329817efc22d652789ff75b594d3c758eb9bc
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>