motion_sense: Trigger interrupt based on AP needs
On Rex, during test, it was observed we would get interrupt every 2
samples.
Setting the interrupt request was delayed beyond a single samples:

Current behavior: need events every 5ms, but the sensor ODR is 208Hz
(4.8ms):
 ts_last_int
 |/event                     /current event           /next event
 |  <------- 4.8ms -------> | <------- 4.8ms -------> |
 + <-------- ec_rate (5ms) ---------->
                            +++ no event generated.   -- event

Fixed behavior, as soon as we are within the area we expect an event, we
allow raising the interrupt: it can happen within 50% of the period of
the last sample, since we may have to set the ODR to a faster frequency
to match the sensor frequency choices.
 ts_last_int
 |/event                     /current event
 |  <------- 4.8ms -------> |
 + <-------- ec_rate (5ms) ---------->
             <--------- time allowed for new interrupt

BUG=b:295069755
TEST=tast run <IP> hardware.SensorIioserviceHard
Unit test: `make -j run-motion_sense_fifo`

Fixes: ff96cd3447b12 ("motion_sense: Trigger AP interrupt based on sensor timestamp")
Change-Id: I10cbe4a0d9b6a35c68a8425617c5c67ff98d944e