ps8805: Handle when the revision register contains 0x4
As talked about in https://crrev.com/c/4939349, it appears that in
some cases we can get PS8805_A2 and PS8805_A3 in a state where the
revision register has 0x40 instead of 0x00/0xa0.

When we see this case, we'll assume A3. Parade has indicated that it's
OK to run the A3 firmware on the A2 stepping. In addition, once we
boot the A2 with the A3 firmware it actually identifies itself as A2
and in the next reboot we'll fix it up.

BUG=b:304772621
TEST=My broken device recovered itself

Change-Id: I7edb7c225274ec5260405b99a929480753fbff25
Signed-off-by: Douglas Anderson <dianders@chromium.org>