servo_updater: --json and --print output for all boards by default
This makes --json and --print behavior consistent with actual
updater behavior.  Previously they would output for servo_v4 by default.

This initial support for printing all boards requires --all to be
specified (if --board= is not).  This requirement is intended to
flush out any poorly written scripts that are relying on --print
implying --board=servo_v4 by default.  This requirement might be
relaxed in the future.

BUG=none
TEST=Tested each of:

$ # both exit non-zero with an error
$ servo_updater --print
$ servo_updater --json

$ # both print info for all servo boards
$ servo_updater --print --all
$ servo_updater --json --all

$ # these print info for the requested boards only
$ servo_updater --print --board=servo_micro
$ servo_updater --json --board=servo_micro
$ servo_updater --print --board=servo_micro --board=c2d2
$ servo_updater --json --board=servo_micro --board=c2d2

Change-Id: I073330ce1d53f25f760a8b6d9d6fb7fd0ef4391a