zephyr: Deduplicate the FWID
The FWID is originally placed in current_image_data.cros_fwid[32].
In CrOS-EC, the FWID fmap section is pointed to the same space.
So other tools can easily read the FWID from the EC image.

But in Zephyr-EC, a separated FWID fmap section is created which
is put at the end of the RO_FW/RW_FW sections. It leads problems:
 * Duplication of the FWID.
 * The size of the RW_FW section (the size to update) is the same
   size as the EC-RW section (the available flash size). It adds
   time penalty to the boot time (calculating the hash) and the
   EC update time.

This CL makes the Zephyr-EC back to the CrOS-EC approach.

BUG=
TEST=

Change-Id: I1e7de14c1d71300b39933cba0a23a6b80972ae8f