rex: Require minimum 45 W charger with low/no battery
The default mimimum charger limit of 15 W is insufficient for booting
rex, so bump up the minimums.

If the battery is missing or low (1% state of charge by default),
require a 45 W or higher charger. We may need to tune this further, but
let's disallow booting on clearly inadequate chargers.

BUG=b:306473685,b:296952944
TEST=booted on rex

Change-Id: I9d0ca14fe20cb7dabc89ed26c92f3f32ef99beb1