zephyr: LED: Add SYSRQ_DEBUG LED alias
In some components (e.g., common/buttons.c), they need to use API
`led_control` to indicate that EC is in certain state, and it is using
the LED ID `SYSRQ_DEBUG`. Not all design has a dedicated LED for
displaying `SYSRQ_DEBUG`, so they would other LED ID to display the same
signal.

Add extra optional DTS properties in the led policy, such that we can
define alias LED ID for overloading that LED as `SYSRQ_DEBUG`, and also
what LED color to display when `SYSRQ_DEBUG` is active.

BUG=b:305864359
TEST=(1) Specify alias pin as battery LED, flash to Starmie device, and
         check the battery LED is flashing when device is in EC debug
         mode.
     (2) /twister -v -i -T zephyr/test/drivers --no-upload-cros-rdb -s
     drivers.led_driver
     (3) /twister -v -i -T zephyr/test/drivers --no-upload-cros-rdb -s
     drivers.led_driver_alias

Change-Id: I5f13da82f7e7c207a29db8c1407df0e8b88a6f43