host: do not abort on persistent flash image read failures
When debugging Cr50 tests, early termination of tests sometimes result
in the creation of an empty persistent flash image file. The following
test runs try to read the image, fail to read the expected number of
files because the file is empty and fire an assert exception of the
test.

Then the user is supposed to find the file and remove it by hand
before the test can run again.

Let's just treat corrupted file the same as the 'no file' case, by
initializing the flash image to all 0xff and reporting the problem
with the image file.

BUG=None
TEST=interrupted tests do not result in the locked system, the test
     can be successfully re-run.

Change-Id: I46efc52de08d69c12d750ade5df7d60348f38a96