helipilot: Bump sha1 of version
The previous release of helipilot did not include the FPC libraries
but creating a new release from the same release branch commit would
result in the same (or very similar) names, which could cause
confusion. This empty commit will modify the version enough to be an
obviously different build.

Unfortunately gerrit doesn't allow submitting CQs with no changed
files. This creates an innocuous change to bump the version.

BUG=b:306401896
TEST=none

Change-Id: I452955690c6fea7f9a9d93f9024d55827aa4482c