zephyr: Restructure CEC DT bindings
Instead of using the same bindings for each CEC ports, add separate
bindings for each driver, along with bindings for a CEC parent node.
This allows each driver to specify its own properties, which will be
used to specify GPIOs for the bitbang driver in the following CL.

Also use a single DTS file for all CEC tests.

BUG=b:287558802
TEST=./twister -s drivers/drivers.cec_bitbang
     ./twister -s drivers/drivers.cec_it83xx

Change-Id: If02923c77b9833093b0058ccd30b65a8a9e3166c