helipilot: Allow helipilot to build from older private repos
The helipilot branches could not be added to the manifest files due to
a repo issue. Workarounds were made to allow building helipilot from
bloonchipper files, but the bloonchipper release is behind
helipilot. This resulted in a dependency issue because newer EC builds
depend upon newer private repo builds and this workaround removes that
dependency. Per b/307922418, this change should be reverted once those
issues are resolved.

BUG=b:306401896
TEST=build helipilot, verify private repos

Cq-Depend: chrome-internal:6606921
Change-Id: I1428d28d96585aed35fe9d7a1a8c40e89290da70