board/hyperdebug: Avoid splitting I2C response into two USB packets
The USB protocol used by uServo for I2C communication is encapsulated in
CMSIS-DAP on HyperDebug, that means a single CMSIS-DAP byte in front of
every request and every response.

The existing code would add the first byte of the response using
`queue_add_unit()`, and a copy of the old I2C logic would then add the
I2C response using `queue_add_units()`.  It turns out that often but not
always, this results in a USB packet with a single byte, followed by
another USB packet with the rest of the response.  While technically
valid, this is undesirable, as it would be more efficient to transmit
the reponse in a single USB packet.

This CL slightly modifies the `usb_i2c_execute()` routine to work
directly on the `rx_buffer` array, making it aware that a single header
byte is sitting at the first position, which should also be transmitted
as part of the response.

BUG=none
TEST=make tast TAST_EXPR=gscdevboard.Ti50Tpm.i2c  (in ti50/common)

Change-Id: I47187a1c5a9f32831dc76804c4ad803c656fd948