fpmcu: Route logs through shell log backend
When the LOG_MODE is IMMEDIATE or DEFERRED, the SHELL_LOG_BACKEND is
enabled automatically. It means that all log messages are printed using
shell instead of `printk()` (when the default MINIMAL log mode is
enabled).

Enabling LOG_PRINTK config causes all `printk()` and `vprintk()` calls
to be routed to logging subsystem.

EC's `CPRINT()` family macros are also using shell subsystem to print
messages.

As a result all messages should be correctly printed.

Please note that, this slightly changes the logs format. It will be
fixed in future using custom formatter.

BUG=b:299924239
TEST=zmake build bloonchipper

Change-Id: I693b5e1c399dcb1a355011bbe42951ccb19752f7