flash_ec: Add apshutdown option
The apshutdown option allows flash_ec to shut down the AP immediately
after flashing is done. When the option is true, flash_ec issues the
apshutdown console command repeatedly 8 times by default. The
apshutdown_try option can be used to change the number of times the
apshutdown command is issued.

The apshutdown_try is set to 16 by default. It has to be large enough
for the shutdown attempts to last beyond a deferred boot triggered
by an RTC reset.

The apshutdown option is off by default.

BUG=b:269804618
TEST=flash_ec --board vell --read /dev/null --apshutdown --apshutdown_try 8

Change-Id: I6182f8bd2648a3ea63b753698fbd7d87001ef0df