flash_ec: Add try_apshutdown option
The try_apshutdown option allows flash_ec to shut down the AP
immediately after flashing is done. This is done by sending the
apshutdown console command repeatedly. The try_apshutdown_count option
can be used to change the number of times the apshutdown command is
issued.

The try_apshutdown_count is set to 16 by default. It has to be large
enough for the shutdown attempts to last beyond a deferred boot
triggered by an RTC reset.

The try_apshutdown option is off by default.

BUG=b:269804618
TEST=Verified a DUT stays in G3 after flashing Vell by:
     flash_ec --board vell --read /dev/null --try_apshutdown \
     --try_apshutdown_count 8

Change-Id: I6182f8bd2648a3ea63b753698fbd7d87001ef0df