board/hyperdebug: Enable PG[15:2] pins early
The STM32L5 chip has 14 pins that run on a second VDDIO2 supply rail.
In order to handle cases where this rail is unpowered, the chip has an
"enable bit", without which it will completely disable the IO circuits
related to those pins.

Since on the HyperDebug board VDDIO2 is hard-wired to the same as the
main supply rail, we should enable that bit, and have done so from
board_init().

However, since the UART console output is among these 14 pins, and since
UART output begins even before entering board_init(), the existing code
cuts off some of the initial boot messages.

This CL moves the code to enable the GPIO circuit into
board_config_pre_init(), so that UART output appears in its entirety.

BUG=none
TEST=watch console output: "--- UART initialized after reboot ---"

Change-Id: I279596d98e84daecbfa2abddf41f290a7515ec0a