tests: exercise smart battery errors with cutoff
This adds a test to exercise the error paths in smart battery accessors
which returns an error if the battery is cut off, which was covered by
happenstance prior to crrev.com/c/4975896 because the battery got stuck
pending cutoff.

The smart battery tests need to be moved to a new suite because other
tests in the drivers.default suite depend on having real battery cutoff
functions and the new test case needs to mock it out.

BUG=b:308185587
TEST=twister

Change-Id: I37c5ca1833e19b06253a021866518496bd8cd192