yavilla: Swap I2C A and C and enable fifo on I2C_SENSOR
Currently there's issue that the latency of sensor
timestamp will be timeout, which leads to fail on
CTS test. We found that this might be related to
I2C_SENSOR transmit with PIO mode, while enabling
FIFO on I2C_SENSOR will increase the speed of I2C
transmission and CTS pass rate. However, there's
issue on fifo enable on I2C C channel (I2C_SENSOR)
(b:290553691),we have to swap I2C A and C channel
to enable fifo on I2C_SENSOR. This CL should be
based on zephyr change:
https://github.com/zephyrproject-rtos/zephyr/pull/64602.

BUG=b:292979502
TEST=make sure CTS test pass.

Change-Id: Ic6c66bfcdde46d4cdbc53cce646af8f31860c2e3