zephyr: Set form factor to convertible for corsola sensor tests
CL:4975165 removes callbacks when disabling gpio interrupts. With this
change, some sensor tests are failing because interrupts are not running
when they are expected to.

Board code will disable motion sense interrupts if the form factor in
fw_config is clamshell (e.g. disable_base_imu_irq() in
steelix/src/board.c). Modify the tests to set the form factor to
convertible.

BUG=b:287558802
TEST=Following tests now pass with CL:4975165:
./twister -s kingler/kingler.alt_sensor
./twister -s kingler/kingler.ponyta
./twister -s krabby/krabby.tentacruel

Change-Id: I7d9890b99070a0e4b5f026d84375e14f576b95b8