move explain_ro_verification_status into gsc-utils
This is an attempt to decouple of ap_ro_verification_status utility
from the Ti50 repo. Three crates have been created:

 - ap_ro_errs: combined AP RO verification errors module with errors
   from some other modules, satisfies all requirements of the
   explain_ro_verification_status application;

 - enum_as: exact copy of the Ti50 crate, can be shared directly.

 - explain_ap_ro_verification_status: the actual application code.

This patch still duplicates a lot of code, and while enum_as crate can
be shared, the error crate is a modification which is harder to share
between the projects.

BUG=b:295109412
TEST=unit tests pass. Compared results of interpreting actual AP RO
     verification statuses by the two versions of the utility, they
     match.

Change-Id: I937743392cf9376e20e3050e7cfbb22ef18d1297
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>