chip/stm32: Properly support clock changes on STM32L5
It seems that when added support for STM32L5 based on L4, I missed an
ifdef related to modifying the clock prescaler of the system
timer (counting microseconds) in response to changes in core clock
frequency.

Prior to this change, the HOOK_TICK which normally happens twice a
second, would happen much less frequently on the HyperDebug board when
switching from pll clock source (104 MHz) to e.g. msi clock source (4
MHz), which would in turn result in watchdog reset soon after the
change.

BUG=none
TEST=clock msi  (observe no longer getting watchdog reset)

Change-Id: I90df155335826cb99e7b6ee924c8d4f86ef5c0fa