board/hyperdebug: Run peripherals at lower clock speed
HyperDebug runs high core clock speed, mostly do be able to perform
logic analyzer functionality using GPIO edge interrupts.

At the same time, we want to be able to run SPI busses at quite low
speed, around 100kHz, in order to be able to potentially capture the SPI
responses of devices under test using the aforementioned logic analyzer
functionality.

The SPI clock is generated by dividing the PCLK by a factor between 2
and 256.  This does not allow slow enough SPI clock, if the PCLK is
identical to the core clock frequency, which exceeds 100MHz.

This CL build on prior EC STM32 support to enable a feature of the STM32
chips to prescale the peripheral clock.

With this change SPI1 and SPI2 can be clocked in any power of 2 between
107kHz and 13.75MHz.  QSPI runs on the higher performance HCLK, and its
clock frequency range remains 430kHz to 110MHz.

BUG=none
TEST=TBD

Change-Id: I2e24a34f1dfe6e1db2c919b9ccf5bded702f8fe0