zephyr: Support deduplicating RO/RW FWID
The EC version is placed in current_image_data.version[32] for each
RO/RW firmware blob.

In CrOS-EC, the FWID fmap section points to the above version[32].
So other tools can easily read the version string from the EC image.

But in Zephyr-EC, a separated FWID fmap section is created which
is put at the end of the RO_FW/RW_FW sections. It leads problems:
 * Duplication of the version string.
 * The size of the RW_FW section (the size to update) is the same
   size as the EC_RW section (the available flash size). It adds
   time penalty to the boot time (calculating the hash) and the
   EC update time.

This CL adds the support of the CrOS-EC way. For projects want to
use it, they can include the `integrated_fwid.dtsi` overlay. The
existing projects, which don't include the overlay, have no change.

BUG=b:198475757
TEST=With the following changes, checked the RO_FRID and RW_FWID fmap
sections point to current_image_data.version[32].

Change-Id: I1e7de14c1d71300b39933cba0a23a6b80972ae8f