chip/stm32: Combine clock-f.h and clock-l4.h
The two files `clock-f.h` and `clock-l4.h` declare some functions
available on some but not all STM32 chips.  Some functions such as
`clock_get_timer_freq()` is in fact available and used on all STM32
chips, another `clock_init()` is also declared in `common/` and need not
be here.

This CL merges the two files into `clock_chip.h`, declaring a set of
clock related STM32 functions, some of which are available on all STM32
variants, others being defined only on a subset of the variants.

BUG=none
TEST=TBD

Change-Id: If30f90c82c2a2b3c346eb7ddceff7a9a28963157