board/hyperdebug: Fix but with SPI1 re-initialization
When SPI1 was added to `spi_devices[]`, I forgot to also add it to the
`spi_device_default_div[]` array used by the "reinit" command to be able
to set the SPI clock speed back to the power-on default value.

This CL changes the logic such that it is no longer necessary to
duplicate the configuration.

BUG=none
TEST=observe "reinit" correctly reset to defaults

Change-Id: Ice9e77883e509e5cf1c1163872b2147fc332b2d7