fingerprint: Disable jtag/swd on boot
There is a bit of a catch'22. We would like to disable JTAG, if the
flash protect is enabled, which means that we need to init flash.
We would like JTAG to be disabled before we initialize the protection
flags.

The alternative method considered was to use the EC gpio module
functionality to enable/disable the debug interface. This was
shown to work in PS8, but it requires that all users declare the
JTAG/SWD pins with the correct default pulls/config in the board
gpio.inc file. It also means that the "disablement" method
may change when another usage of GPIO_ALT_FUNC_NONE comes along.
Furthermore, this method would be less portable to Zephyr.

Critical: You can't configure the port A and B gpio module unless
the clock are enable/configured, thus disbling the pins as the
first step of main() is not possible.

BUG=b:142556193
TEST=make proj-dartmonkey proj-bloonchipper
TEST=make BOARD=bloonchipper -j test-debug

Change-Id: Ia51d7b53734dfe36c9e4b9c116f16d94f6980496