fingerprint: Disable JTAG/SWD on boot
There is a bit of a catch'22. We would like to disable JTAG, if the
flash protect is enabled, which means that we need to init flash.
We would like JTAG to be disabled before we initialize the protection
flags.

The alternative method considered was to use the EC gpio module
functionality to enable/disable the debug interface. This was
shown to work in PS8, but it requires that all users declare the
JTAG/SWD pins with the correct default pulls/config in the board
gpio.inc file. It also means that the "disablement" method
may change when another usage of GPIO_ALT_FUNC_NONE comes along.
Furthermore, this method would be less portable to Zephyr.

Critical: You can't configure the port A and B gpio module unless
the clock are enable/configured, thus disbling the pins as the
first step of main() is not possible.

Calling gpio_config_module(MODULE_DEBUG, 0) would work for disabling
the debug interface, but the following are some reasons why we avoid this:
- For security, if we want the compiler/linker to remove the enable
  code path, we can't call a singular function with a one bit difference
- The common module disable function doesn't change both the mode and
  AF selector from the defaults, when disabled. It only changes the mode.
- We would need to make more changes to make the common module disable
  code path (the GPIO_ALT_FUNC_NONE path) sets the pins to analog,
  instead of digital input.

BUG=b:142556193
TEST=make proj-dartmonkey proj-bloonchipper
TEST=make BOARD=bloonchipper -j test-debug
TEST=# Add a pause for manual user power cycling to test script.
     # Unplug debugger during pause, too.
     ./test/run_device_tests.py --board bloonchipper --tests debug_stm32
TEST=# Add a pause for manual user power cycling to test script.
     # Unplug debugger during pause, too.
     ./test/run_device_tests.py --board dartmonkey --tests debug_stm32

Change-Id: Ia51d7b53734dfe36c9e4b9c116f16d94f6980496