board/hyperdebug: Allow QSPI from alternative clock
HyperDebug runs high core clock speed, mostly do be able to perform
logic analyzer functionality using GPIO edge interrupts.

At the same time, we want to be able to run SPI busses at quite low
speed, around 100kHz, in order to be able to potentially capture the SPI
responses of devices under test using the aforementioned logic analyzer
functionality.

By default, the QSPI controller divides the core clock by an integer in
the range 1-256, which in our case results in a SPI bus clock in the
range 110MHz - 430kHz.  Not quite slow enough to allow e.g. capturing of
SPI passthrough on the OpenTitan chip.

Fortunately, the QSPI controller can alternatively use the MSI clock of
STM32, and it happens that HyperDebug does not use the MSI clock for
other things, so we can change the MSI clock frequency to any of its 12
possible values in the range 100kHz - 48MHz, and use that is input to the
QSPI prescaler.  This greatly increases the possible SPI clock speeds.

This CL adds logic to compute the best approximation to the requested
frequency, using either the 110MHz system clock, or each of the 12
possible MSI clock speeds.

Change-Id: I2e24a34f1dfe6e1db2c919b9ccf5bded702f8fe0