chip/stm32: STM32L5: 8x / 16x oversampling on console UART
There is already support in usart-stm32l5.c for putting UARTS into x8 or
x16 oversampling mode depending on baud rate.  This applies to uarts
used for forwarding via USB.

The console UART is handled in uart.c, which is shared between all the
STM32 families.  Due to an omission, the oversampling logic was not
enabled when L5 support was added.

BUG=none
TEST=boot hyperdebug and observe serial console

Change-Id: I32a31b638cb842b2e31cf275148759162e5c0dd5