chip/stm32: Run UARTS on STM32L5 from 16Mhz MSI clock
Prior to this change, the code in uart_init() in uart.c for the L4
familiy was strange, it would always set UART1 to use 16MHz MSI and set
LPUART to use SYSCLK (despite uart_freq_change() expecting 16MHz no
matter the UART).  This would work only when UARTN is defined as 1.

This CL does not change the logic for G4, but separates the logic from
L4, and then changes the L4 logic to set only the actual UART indicated
by UARTN to use 16MHz MSI.  This same logic is put in place for L5.

Also, this CL changes usart-stm32l5.c (the file dealing with non-console
UARTs on L5) to also set any such UARTs to use 16MHz MSI, eliminating
the need for re-computing divisor on changes to core clock frequency.

BUG=none
TEST=boot hyperdebug, switch clock, observe no interruption on console

Change-Id: I1e0d360a526bd9c115ea109f16322f14c9b1ce51